+++ grizzly/steps/scenario/response.py
@@ -8,8 +8,8 @@ from behave import register_type, when, then  # pylint: disable=no-name-in-modul
 
 from ...context import GrizzlyContext
 from ...task import RequestTask
-from ...types import ResponseContentType, ResponseTarget
-from ...utils import add_save_handler, add_validation_handler, add_request_task_response_status_codes
+from ...types import ResponseContentType, ResponseTarget, str_response_content_type
+from ..helpers import add_save_handler, add_validation_handler, add_request_task_response_status_codes
 
 
 @parse.with_pattern(r'is( not)?', regex_group_count=1)
@@ -29,21 +29,10 @@ def parse_response_target(text: str) -> ResponseTarget:
         raise ValueError(f'"{text}" is an unknown response target')
 
 
-def parse_response_content_type(text: str) -> ResponseContentType:
-    if text.strip() in ['application/json', 'json']:
-        return ResponseContentType.JSON
-    elif text.strip() in ['application/xml', 'xml']:
-        return ResponseContentType.XML
-    elif text.strip() in ['text/plain', 'plain']:
-        return ResponseContentType.PLAIN
-    else:
-        raise ValueError(f'"{text}" is an unknown response content type')
-
-
 register_type(
     Condition=parse_condition,
     ResponseTarget=parse_response_target,
-    ResponseContentType=parse_response_content_type,
+    ResponseContentType=str_response_content_type,
 )
