+++ grizzly/steps/scenario/setup.py
@@ -7,8 +7,7 @@ from behave.runner import Context
 from behave import register_type, given  # pylint: disable=no-name-in-module
 
 from ...context import GrizzlyContext
-from ...testdata.utils import merge_dicts
-from ...utils import create_context_variable, resolve_variable
+from ...utils import create_context_variable, resolve_variable, merge_dicts
 
 
 @parse.with_pattern(r'(iteration[s]?)')
