+++ grizzly/steps/scenario/tasks.py
@@ -4,7 +4,7 @@ from typing import cast
 from behave.runner import Context
 from behave import register_type, then  # pylint: disable=no-name-in-module
 
-from ...utils import add_request_task
+from ..helpers import add_request_task
 from ...types import RequestDirection, RequestMethod
 from ...context import GrizzlyContext
 from ...task import PrintTask, SleepTask
