+++ grizzly/types.py
@@ -64,6 +64,63 @@ class RequestMethod(Enum, AdvancedEnum, settings=NoAlias):
     def direction(self) -> RequestDirection:
         return self.value
 
+class TemplateData(dict):
+    @classmethod
+    def guess_datatype(cls, value: Any) -> 'TemplateDataType':
+        if isinstance(value, (int, bool, float)):
+            return value
+
+        check_value = value.replace('.', '', 1)
+        casted_value: 'TemplateDataType'
+
+        if check_value[0] == '-':
+            check_value = check_value[1:]
+
+        if check_value.isdecimal():
+            if float(value) % 1 == 0:
+                if value.startswith('0'):
+                    casted_value = str(value)
+                else:
+                    casted_value = int(float(value))
+            else:
+                casted_value = float(value)
+        elif value.lower() in ['true', 'false']:
+            casted_value = value.lower() == 'true'
+        else:
+            casted_value = str(value)
+            if casted_value[0] in ['"', "'"]:
+                if casted_value[0] != casted_value[-1] and casted_value.count(casted_value[0]) % 2 != 0:
+                    raise ValueError(f'{value} is incorrectly quoted')
+
+                if casted_value[0] == casted_value[-1]:
+                    casted_value = casted_value[1:-1]
+            elif casted_value[-1] in ['"', "'"] and casted_value[-1] != casted_value[0] and casted_value.count(casted_value[-1]) % 2 != 0:
+                    raise ValueError(f'{value} is incorrectly quoted')
+
+        return casted_value
+
+    def __setitem__(self, key: str, value: 'TemplateDataType') -> None:
+        caster: Optional[Callable] = None
+
+        if '.' in key:
+            [name, _] = key.split('.', 1)
+            try:
+                from .testdata.variables import load_variable
+                variable = load_variable(name)
+                caster = variable.__base_type__
+            except AttributeError:
+                pass
+
+        if isinstance(value, str):
+            if caster is None:
+                value = self.guess_datatype(value)
+            else:
+                value = caster(value)
+        elif caster is not None:
+            value = caster(value)
+
+        super().__setitem__(key, value)
+
 
 HandlerType = Callable[[Tuple[ResponseContentType, Any], User, Optional[ResponseContextManager]], None]
 
@@ -85,3 +142,14 @@ def bool_typed(value: str) -> bool:
 
 def int_rounded_float_typed(value: str) -> int:
     return int(round(float(value)))
+
+
+def str_response_content_type(value: str) -> ResponseContentType:
+    if value.strip() in ['application/json', 'json']:
+        return ResponseContentType.JSON
+    elif value.strip() in ['application/xml', 'xml']:
+        return ResponseContentType.XML
+    elif value.strip() in ['text/plain', 'plain']:
+        return ResponseContentType.PLAIN
+    else:
+        raise ValueError(f'"{value}" is an unknown response content type')
