+++ grizzly/users/sftp.py
@@ -31,7 +31,7 @@ from os import path, environ, mkdir
 from locust.exception import StopUser
 
 from .meta import ContextVariables, FileRequests
-from ..testdata.utils import merge_dicts
+from ..utils import merge_dicts
 from ..clients import SftpClientSession
 from ..types import RequestMethod
 from ..task import RequestTask
