+++ grizzly/utils.py
@@ -1,35 +1,26 @@
 import logging
 import os
-import re
-import json
 
 
 from typing import Callable, Generic, TypeVar, Type, List, Any, Dict, Tuple, Optional, cast, Generator
 from types import FunctionType
 from importlib import import_module
 from functools import wraps
-from urllib.parse import urlparse
 from contextlib import contextmanager
-
-import jinja2 as j2
+from collections import namedtuple
+from collections.abc import Mapping
+from copy import deepcopy
 
 from behave.runner import Context
-from behave.model import Scenario, Row
+from behave.model import Scenario
 from behave.model_core import Status
-from locust.clients import ResponseContextManager
 from locust.user.users import User
 from locust import TaskSet, between
 from jinja2 import Template, Environment
 from jinja2.meta import find_undeclared_variables
 
-from .testdata.utils import transform, merge_dicts
-from .testdata.models import TemplateData, TemplateDataType
-from .users.meta import ContextVariables
-from .exceptions import ResponseHandlerError
-from .types import HandlerType, ResponseContentType, RequestMethod, ResponseTarget, ResponseAction
-from .transformer import PlainTransformer, transformer
+from .types import TemplateData, TemplateDataType
 from .context import GrizzlyContext, GrizzlyContextScenario
-from .task import RequestTask
 
 
 logger = logging.getLogger(__name__)
@@ -83,368 +74,6 @@ class catch:
         return cast(WrappedFunc, wrapper)
 
 
-def create_request_task(context: Context, method: RequestMethod, source: str, endpoint: str, name: Optional[str] = None) -> RequestTask:
-    grizzly = cast(GrizzlyContext, context.grizzly)
-    request = _create_request_task(context.config.base_dir, method, source, endpoint, name)
-    request.scenario = grizzly.scenario
-
-    return request
-
-
-def _create_request_task(base_dir: str, method: RequestMethod, source: str, endpoint: str, name: Optional[str] = None) -> RequestTask:
-    path = os.path.join(base_dir, 'requests')
-    j2env = j2.Environment(
-        autoescape=False,
-        loader=j2.FileSystemLoader(path),
-    )
-
-    template: Optional[j2.Template] = None
-
-    try:
-        if source is not None:
-            possible_file = os.path.join(path, source)
-            # minify json files, to increase performance when jinja creates template
-            if os.path.isfile(possible_file):
-                with open(possible_file, 'r') as fd:
-                    try:
-                        source = json.dumps(json.load(fd))
-                        raise RuntimeError()
-                    except json.decoder.JSONDecodeError:
-                        # not json contents, so do not minify
-                        pass
-
-            template = j2env.get_template(source)
-            if name is None:
-                name = source.replace('.j2.json', '')
-
-            with open(os.path.join(path, source), 'r') as fd:
-                source = fd.read()
-    except (j2.exceptions.TemplateNotFound, RuntimeError):
-        if source is not None:
-            template = j2env.from_string(source)
-
-        if name is None:
-            name = '<unknown>'
-
-    request = RequestTask(method, name=name, endpoint=endpoint)
-    request.template = template
-    request.source = source
-
-    return request
-
-
-def add_request_task_response_status_codes(request: RequestTask, status_list: str) -> None:
-    for status in status_list.split(','):
-        request.response.add_status_code(int(status.strip()))
-
-
-def add_request_task(context: Context, method: RequestMethod, source: str, name: Optional[str] = None, endpoint: Optional[str] = None) -> None:
-    grizzly = cast(GrizzlyContext, context.grizzly)
-    scenario_tasks_count = len(grizzly.scenario.tasks)
-
-    table: List[Optional[Row]]
-
-    if context.table is not None:
-        table = context.table
-    else:
-        table = [None]
-
-    for row in table:
-        if endpoint is None:
-            if scenario_tasks_count == 0:
-                raise ValueError(f'no endpoint specified')
-
-            last_request = grizzly.scenario.tasks[-1]
-
-            if not isinstance(last_request, RequestTask):
-                raise ValueError('previous task was not a request')
-
-            if last_request.method != method:
-                raise ValueError(f'can not use endpoint from previous request, it has different method')
-
-            endpoint = last_request.endpoint
-        else:
-            parsed = urlparse(endpoint)
-            if len(parsed.netloc) > 0:
-                raise ValueError(f'endpoints should only contain path relative to {grizzly.scenario.context["host"]}')
-
-        orig_endpoint = endpoint
-        orig_name = name
-        orig_source = source
-
-        if row is not None:
-            for key, value in row.as_dict().items():
-                endpoint = endpoint.replace(f'{{{{ {key} }}}}', value)
-                if name is not None:
-                    name = name.replace(f'{{{{ {key} }}}}', value)
-                if source is not None:
-                    source = source.replace(f'{{{{ {key} }}}}', value)
-
-        request_task = create_request_task(context, method, source, endpoint, name)
-
-        endpoint = orig_endpoint
-        name = orig_name
-        source = orig_source
-
-        grizzly.scenario.tasks.append(request_task)
-
-
-def get_matches(
-    input_get_values: Callable[[Any], List[str]],
-    match_get_values: Callable[[Any], List[str]],
-    input_payload: Any,
-) -> Tuple[List[Any], List[Any]]:
-    '''Find all values in `input_context`.
-
-    Args:
-        input_get_values (Callable[[Any], List[str]]): function that returns all values matching `expression`
-        input_match_values (Callable[[Any], List[str]]): function that checks that a value has correct value
-        input_context (Tuple[ResponseContentType, Any]): content type and transformed payload
-
-    Returns:
-        Tuple[List[Any], List[Any]]: list of all values and list of all matched values of those
-    '''
-    values = input_get_values(input_payload)
-
-    # get a list of all matches in values
-    matches: List[str] = []
-    for value in values:
-        matched_values = match_get_values(value)
-
-        if len(matched_values) < 1:
-            continue
-
-        matched_value = matched_values[0]
-
-        if matched_value is None or len(matched_value) < 1:
-            continue
-
-        matches.append(matched_value)
-
-    return values, matches
-
-
-def handler_logic(
-    input_context: Tuple[ResponseContentType, Any],
-    expression: str,
-    match_with: str,
-    user: ContextVariables,
-    callback: Callable[[str, Optional[Any]], None],
-    condition: bool,
-) -> None:
-    '''Contains common logic for both save and validation handlers.
-
-    Args:
-        input_context (Tuple[ResponseContentType, Any]): content type and transformed payload
-        expression (str): expression to extract value from `input_context`
-        match_with (str): regular expression that the extracted value must match
-        user (ContextVariablesUser): user that executed task (request)
-        callback (Callable[[str, Optional[Any]], None]): specific logic for either save or validation handler
-        condition (bool): used by validation handler for negative matching
-    '''
-    input_content_type, input_payload = input_context
-    interpolated_expression = j2.Template(expression).render(user.context_variables)
-    interpolated_match_with = j2.Template(match_with).render(user.context_variables)
-
-    transform = transformer.available.get(input_content_type, None)
-    if transform is None:
-        raise TypeError(f'could not find a transformer for {input_content_type.name}')
-
-    if not transform.validate(interpolated_expression):
-        raise TypeError(f'"{interpolated_expression}" is not a valid expression for {input_content_type.name}')
-
-    input_get_values = transform.parser(interpolated_expression)
-    match_get_values = PlainTransformer.parser(interpolated_match_with)
-
-    values, matches = get_matches(
-        input_get_values,
-        match_get_values,
-        input_payload,
-    )
-
-    number_of_matches = len(matches)
-
-    if number_of_matches != 1:
-        if number_of_matches < 1 and not condition:
-            logger.error(f'"{interpolated_expression}": "{interpolated_match_with}" not in "{values}"')
-        elif number_of_matches > 1:
-            logger.error(f'"{interpolated_expression}": "{interpolated_match_with}" has multiple matches in "{values}"')
-
-        match = None
-    else:
-        match = matches[0]
-
-    callback(interpolated_expression, match)
-
-
-def generate_validation_handler(expression: str, match_with: str, condition: bool) -> HandlerType:
-    '''Generates a handler that will validate a value from an input.
-
-    Args:
-        expression (str): how to find the specified value, can contain templating variables
-        match_with (str): regular expression that any result from `expression` must match
-        condition (bool): if the match should or should not match
-
-    Returns:
-        HandlerType: function that will validate values in a response during runtime
-    '''
-    def validate(
-        input_context: Tuple[ResponseContentType, Any],
-        user: ContextVariables,
-        response: Optional[ResponseContextManager] = None,
-    ) -> None:
-        '''Actual handler that will run after a response has been received by an task.
-
-        Args:
-            input_context (Tuple[ResponseContentType, Any]): content type and transformed payload
-            user (ContextVariablesUser): user that executed task (request)
-            response (Optional[ResponseContextManager]): optional response context, only if `user` does HTTP requests
-        '''
-        def callback(
-            interpolated_expression: str,
-            match: Optional[Any],
-        ) -> None:
-            '''Validation specific logic that will handle the `match` based on `expression` and `match_with`.
-
-            Args:
-                interpolated_expression (str): `expression` with templating variables resolved
-                match (Optional[Any]): value based on `expression` that matches `match_with`
-            '''
-            result = match is not None if condition == True else match is None
-
-            if result:
-                message = f'"{interpolated_expression}": "{match_with} was {match}"'
-                if response is not None:
-                    response.failure(message)
-                else:
-                    raise ResponseHandlerError(message)
-
-        handler_logic(input_context, expression, match_with, user, callback, condition)
-
-    return validate
-
-
-def generate_save_handler(expression: str, match_with: str, variable: str) -> HandlerType:
-    '''Generates a handler that will extract a value from the response and save it in a templating variable.
-
-    Args:
-        expression (str): how to find the specified value, can contain templating variables
-        match_with (str): regular expression that any result from `expression` must match
-        variable (str): name of templating variable in the user context
-
-    Returns:
-        HandlerType: function that will save values from responses during run time
-    '''
-    def save(
-        input_context: Tuple[ResponseContentType, Any],
-        user: ContextVariables,
-        response: Optional[ResponseContextManager] = None,
-    ) -> None:
-        '''Actual handler that will run after a response has been received by an task.
-
-        Args:
-            input_context (Tuple[ResponseContentType, Any]): content type and transformed payload
-            user (ContextVariablesUser): user that executed task (request)
-            response (Optional[ResponseContextManager]): optional response context, only if `user` does HTTP requests
-        '''
-        def callback(
-            interpolated_expression: str,
-            match: Optional[Any],
-        ) -> None:
-            '''Validation specific logic that will handle the `match` based on `expression` and `match_with`.
-
-            Args:
-                interpolated_expression (str): `expression` with templating variables resolved
-                match (Optional[Any]): value based on `expression` that matches `match_with`
-            '''
-            user.set_context_variable(variable, match)
-
-            if match is None:
-                message = f'"{interpolated_expression}" did not match value'
-                if response is not None:
-                    response.failure(message)
-                else:
-                    raise ResponseHandlerError(message)
-
-        handler_logic(input_context, expression, match_with, user, callback, False)
-
-    return save
-
-
-def _add_response_handler(
-    context: GrizzlyContext,
-    target: ResponseTarget,
-    action: ResponseAction,
-    expression: str,
-    match_with: str,
-    variable: Optional[str] = None,
-    condition: Optional[bool] = None,
-) -> None:
-    scenario_tasks_count = len(context.scenario.tasks)
-
-    if variable is not None and variable not in context.state.variables:
-        raise ValueError(f'variable {variable} has not been declared')
-
-    if not scenario_tasks_count > 0:
-        raise ValueError('no request source has been added!')
-
-    if len(expression) < 1:
-        raise ValueError('expression is empty')
-
-    # latest request
-    request = context.scenario.tasks[-1]
-
-    if not isinstance(request, RequestTask):
-        raise ValueError('latest task was not a request')
-
-    if '{{' in match_with and '}}' in match_with:
-        context.scenario.orphan_templates.append(match_with)
-
-    if '{{' in expression and '}}' in match_with:
-        context.scenario.orphan_templates.append(expression)
-
-    if action == ResponseAction.SAVE:
-        if variable is None:
-            raise ValueError('variable is not set')
-
-        handler = generate_save_handler(expression, match_with, variable)
-    elif action == ResponseAction.VALIDATE:
-        if condition is None:
-            raise ValueError('condition is not set')
-
-        handler = generate_validation_handler(expression, match_with, condition)
-
-    if target == ResponseTarget.METADATA:
-        add_listener = request.response.handlers.add_metadata
-    elif target == ResponseTarget.PAYLOAD:
-        add_listener = request.response.handlers.add_payload
-
-    add_listener(handler)
-
-
-def add_save_handler(context: GrizzlyContext, target: ResponseTarget, expression: str, match_with: str, variable: str) -> None:
-    _add_response_handler(context, target, ResponseAction.SAVE, expression=expression, match_with=match_with, variable=variable)
-
-
-def add_validation_handler(context: GrizzlyContext, target: ResponseTarget, expression: str, match_with: str, condition: bool) -> None:
-    _add_response_handler(context, target, ResponseAction.VALIDATE, expression=expression, match_with=match_with, condition=condition)
-
-
-def normalize_step_name(step_name: str) -> str:
-    return re.sub(r'"[^"]*"', '""', step_name)
-
-
-def in_correct_section(func: FunctionType, expected: List[str]) -> bool:
-    try:
-        actual = '.'.join(func.__module__.rsplit('.', 1)[:-1])
-    except AttributeError:  # function does not belong to a module
-        actual = 'custom'
-
-    return (
-        actual.startswith('grizzly.') and actual in expected
-    ) or not actual.startswith('grizzly.')
-
-
 @contextmanager
 def fail_direct(context: Context) -> Generator[None, None, None]:
     # save original values
@@ -479,6 +108,8 @@ def create_user_class_type(scenario: GrizzlyContextScenario, global_context: Opt
 
     context: Dict[str, Any] = {}
     contexts: List[Dict[str, Any]] = []
+
+    from .users.meta import ContextVariables
 
     if issubclass(base_user_class_type, ContextVariables):
         contexts.append(base_user_class_type._context)
@@ -546,3 +177,66 @@ def resolve_variable(grizzly: GrizzlyContext, value: str, guess_datatype: Option
         resolved_variable = TemplateData.guess_datatype(resolved_variable)
 
     return resolved_variable
+
+
+def merge_dicts(merged: Dict[str, Any], source: Dict[str, Any]) -> Dict[str, Any]:
+    merged = deepcopy(merged)
+    source = deepcopy(source)
+
+    for k in source.keys():
+        if (k in merged and isinstance(merged[k], dict)
+                and isinstance(source[k], Mapping)):
+            merged[k] = merge_dicts(merged[k], source[k])
+        else:
+            merged[k] = source[k]
+
+    return merged
+
+
+def transform(data: Dict[str, Any], raw: Optional[bool] = False) -> Dict[str, Any]:
+    testdata: Dict[str, Any] = {}
+
+    for key, value in data.items():
+        if '.' in key:
+            paths: List[str] = key.split('.')
+            variable = paths.pop(0)
+            path = paths.pop()
+            struct = {path: value}
+            paths.reverse()
+
+            for path in paths:
+                struct = {path: {**struct}}
+
+            if variable in testdata:
+                testdata[variable] = merge_dicts(testdata[variable], struct)
+            else:
+                testdata[variable] = {**struct}
+        else:
+            testdata[key] = value
+
+    if not raw:
+        return _objectify(testdata)
+    else:
+        return testdata
+
+
+def _objectify(testdata: Dict[str, Any]) -> Dict[str, Any]:
+    for variable, attributes in testdata.items():
+        if not isinstance(attributes, dict):
+            continue
+
+        attributes = _objectify(attributes)
+        testdata[variable] = namedtuple('Testdata', attributes.keys())(**attributes)
+
+    return testdata
+
+
+def in_correct_section(func: FunctionType, expected: List[str]) -> bool:
+    try:
+        actual = '.'.join(func.__module__.rsplit('.', 1)[:-1])
+    except AttributeError:  # function does not belong to a module
+        actual = 'custom'
+
+    return (
+        actual.startswith('grizzly.') and actual in expected
+    ) or not actual.startswith('grizzly.')
